chore: move hypothesis_cleanup to conftest
