+++ packages/valory/skills/abstract_round_abci/tests/conftest.py
@@ -21,11 +21,17 @@
 
 # pylint: skip-file
 
-from typing import Dict
+import shutil
+from contextlib import suppress
+from pathlib import Path
+from typing import Dict, Generator
 
 import pytest
 
 from packages.valory.skills.abstract_round_abci.io_.store import StoredJSONType
+
+
+PACKAGE_DIR = Path(__file__).parent.parent
 
 
 @pytest.fixture
@@ -38,3 +44,13 @@ def dummy_obj() -> StoredJSONType:
 def dummy_multiple_obj(dummy_obj: StoredJSONType) -> Dict[str, StoredJSONType]:
     """Many dummy custom objects to test the storing with."""
     return {f"test_obj_{i}": dummy_obj for i in range(10)}
+
+
+@pytest.fixture(scope="session", autouse=True)
+def hypothesis_cleanup() -> Generator:
+    """Fixture to remove hypothesis directory after tests."""
+    yield
+    hypothesis_dir = PACKAGE_DIR / ".hypothesis"
+    if hypothesis_dir.exists():
+        with suppress(OSError, PermissionError):
+            shutil.rmtree(hypothesis_dir)
